+++ generator/templates/services/commercetools/stores.tf
@@ -1,20 +1,20 @@
-{% for store in commercetools.managed_stores %}
-resource "commercetools_store" "{{ store.key }}" {
-  key  = "{{ store.key }}"
+{% for store in commercetools.ManagedStores %}
+resource "commercetools_store" "{{ store.Key }}" {
+  key  = "{{ store.Key }}"
   name = {
-    {% for language, localized_name in store.name.items() %}
-        {{ language  }} = "{{ localized_name }}"
+    {% for language, localizedName in store.Name %}
+        {{ language  }} = {{ localizedName|tf }}
     {% endfor %}
   }
-  {% if store.languages %}
-  languages  = [{% for language in store.languages %}"{{ language }}"{% if not loop.last %},{% endif %}{% endfor %}]
+  {% if store.Languages %}
+  languages  = [{% for language in store.Languages %}"{{ language }}"{% if not loop.last %},{% endif %}{% endfor %}]
   {% endif %}
 
-  {% if store.distribution_channels %}
-  distribution_channels = [{% for dc in store.distribution_channels %}commercetools_channel.{{ dc }}.key{% if not loop.last %},{% endif %}{% endfor %}]
+  {% if store.DistributionChannels %}
+  distribution_channels = [{% for dc in store.DistributionChannels %}commercetools_channel.{{ dc }}.key{% if not loop.last %},{% endif %}{% endfor %}]
   {% endif %}
-  {% if store.supply_channels %}
-  supply_channels = [{% for sc in store.supply_channels %}commercetools_channel.{{ sc }}.key{% if not loop.last %},{% endif %}{% endfor %}]
+  {% if store.SupplyChannels %}
+  supply_channels = [{% for sc in store.SupplyChannels %}commercetools_channel.{{ sc }}.key{% if not loop.last %},{% endif %}{% endfor %}]
   {% endif %}
 }
 {% endfor %}